0.2.6.1-dev: bumped version Adjusted settings options so that when selecting a data league it will automatically check if ssf and then set a poeladder setting, same with checking/unchecking ssf button, these still require refreshing the tree/UI
Adjusted so that the True/False state of "Picked" is actually saved/rendered

Simplified and merged the loading data logic, moved the ladder_identifier to be well the key of the dataset instead of the "league" variable so that we can use the constructed poeladder_data_league var.

Updated a few cases of logging so that they display the actual league used and their usage, i.e poeladder or poe.ninja, added owned / total in the initializing of the tool/changing data leagues.

Fixed duplicate checking using the wrong term (i.e was checking `Ambu's` if it matches with `Ambu` which well would result in allowing duplicates).

Removed "\n=== Fetching items for {league} ===" as we are already displaying the number of items per league type for poe.ninja